+++ app/common/services/datapackage/index.js
@@ -1,10 +1,12 @@
 import angular from 'angular';
-import {DataService} from './dataservice';
+
 import addDataPackageResolver from './resolver';
+import {run, dataservice} from './dataservice';
 
 const module = angular
   .module('projectX.dataService', ['ngRoute'])
-  .service('dataService', DataService)
+  .service('dataService', dataservice)
+  .run(run)
   .decorator('$route', ['$delegate', $delegate => {
     angular.forEach($delegate.routes, addDataPackageResolver);
     return $delegate;