+++ app/common/services/datapackage/mime.js
@@ -1,36 +1,6 @@
-import MimeLookup from 'mime-lookup';
+// for 1.0.0-rc.3 compatability
 
-// mime types supported by dataservices
-export const MimeDb = {
-  'text/tab-separated-values': {
-    source: 'iana',
-    compressible: true,
-    extensions: ['tsv']
-  },
-  'text/csv': {
-    source: 'iana',
-    compressible: true,
-    extensions: ['csv']
-  },
-  'text/plain': {
-    source: 'iana',
-    compressible: true,
-    extensions: ['txt', 'text', 'conf', 'def', 'list', 'log', 'in', 'ini']
-  },
-  'text/yaml': {
-    extensions: ['yaml', 'yml']
-  },
-  'text/matrix': {
-    source: 'custom',
-    extensions: ['matrix', 'expression', 'osc']
-  },
-  'application/json': {
-    source: 'iana',
-    charset: 'UTF-8',
-    compressible: true,
-    extensions: ['json', 'map', 'json5']
-  }
-};
+import dp from './datapackage';
 
-export const mime = new MimeLookup(MimeDb);
+export const mime = dp.processor.mime;
 export default mime;