+++ jspm.browser.js
@@ -2,6 +2,7 @@ SystemJS.config({
   production: true,
   paths: {
     "github:": "jspm_packages/github/",
-    "npm:": "jspm_packages/npm/"
+    "npm:": "jspm_packages/npm/",
+    "project-chi/": "src/"
   }
 });